# redux/post-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/posts/{id}":{"put":{"tags":["post-controller"],"operationId":"updatePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Post"}}}}}},"delete":{"tags":["post-controller"],"operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/posts/new":{"post":{"tags":["post-controller"],"operationId":"createPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Post"}}}}}}},"/posts/{username}":{"get":{"tags":["post-controller"],"operationId":"getPostsByUsername","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Post"}}}}}}}},"/posts/public_feed":{"get":{"tags":["post-controller"],"operationId":"getPublicFeed","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Post"}}}}}}}},"/posts/all/{id}":{"get":{"tags":["post-controller"],"operationId":"getPostsByUserId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Post"}}}}}}}}},"components":{"schemas":{"Post":{"required":["pipelineId","text","userId","visibility"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"image":{"type":"string"},"text":{"type":"string"},"pipelineId":{"type":"integer","format":"int64"},"visibility":{"type":"string"},"userId":{"type":"integer","format":"int64"}}}}}}